primera/requirements.txt
pytorch==1.10.0
pytorch-lightning==1.7.4
datasets==2.2.2
nltk==3.7
tqdm==4.64.0
rouge-score==0.0.4
transformers==4.19.2
wandb==0.12.21